src/api_gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
pydantic
pydantic-settings
PyJWT==2.8.0
